backend/pyproject.toml
[project]
name = "nbim-dividend-reconciliation"
version = "0.1.0"
description = "NBIM dividend reconciliation system using AI for break detection"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.1.4",
    "anthropic>=0.7.8",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []
